Add repl subcommand
